requirements.txt
selenium==4.15.2
python-dotenv==1.0.0
schedule==1.2.0